requirements.txt
transformers==4.27.4
sentencepiece==0.1.97
sacremoses==0.0.53
torch==2.0.0
gradio==3.24.1
pillow>=9.0.0
piexif>=1.1.3
onnxruntime>=1.12.0
protobuf<=3.20.1,>=3.12.2
opencv-python==4.7.0.72
huggingface-hub==0.13.2
clip-interrogator==0.6.0
